+++ lib/src/_columns.dart
@@ -4,8 +4,13 @@ import '_modifiers.dart';
 
 // https://bulma.io/documentation/columns/basics/
 
-Element columns({Mods mods, content}) =>
-    apply(new Element('div', classes: ['columns'], content: content), mods);
+Element columns({Mods mods, content, bool isMultiline, bool isCentered}) =>
+    apply(
+        new Element('div', content: content)
+          ..addClass('columns')
+          ..addClass(multilineClass(isMultiline))
+          ..addClass(centeredClass(isCentered)),
+        mods);
 
 Element column({Mods mods, content}) =>
     apply(new Element('div', classes: ['column'], content: content), mods);
@@ -15,4 +20,3 @@ Element column({Mods mods, content}) =>
 // TODO: add Responsiveness
 // TODO: add Nesting
 // TODO: add Gap
-// TODO: add Options