+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Removed `BulmaComponent`, clients should use `Mods` instead.
 - Removed most `Component` classes, using simple methods instead in simple cases.
 
+New features:
+
+- Implemented column options.
+
 ## 0.1.0
 
 - `Input.value` to provide the current value in the `<input>` field.